Abandoned liblogiops, moved code to logictl
The DBus client is easy enough to implement, creating a library is
unnecessary and causes more issues.
