+++ src/logictl/logictl.cpp
@@ -1,4 +1,4 @@
-#include <liblogiops_client/Control.h>
+#include "client/Control.h"
 #include <stdio.h>
 #include <cstring>
